requirements.txt
torch>=1.9.0
transformers>=4.5.0
pandas>=1.3.0
numpy>=1.19.0
scikit-learn>=0.24.0
matplotlib>=3.4.0
seaborn>=0.11.0
umap-learn>=0.5.0
hdbscan>=0.8.0
nltk>=3.6.0 